Combine "geometry" and "globals" into "variables"
