+++ packages/shader-composer/src/stdlib/index.ts
@@ -4,8 +4,6 @@
 
 export * from "./artistic"
 export * from "./attributes"
-export * from "./geometry"
-export * from "./globals"
 export * from "./logic"
 export * from "./masters"
 export * from "./math"
@@ -15,4 +13,5 @@ export * from "./spaces"
 export * from "./textures"
 export * from "./uniforms"
 export * from "./values"
+export * from "./variables"
 export * from "./vectors"
