+++ packages/shader-composer/src/stdlib/variables.ts
@@ -1,12 +1,38 @@
-/** @packageDocumentation
- * Geometryyyy
- */
-
 import { Vector2 } from "three"
 import { $, Expression } from "../expressions"
 import { injectAPI, Input } from "../units"
 import { $localToViewSpace, $localToWorldSpace } from "./spaces"
-import { Bool, Mat4, Vec2, Vec3 } from "./values"
+import { Bool, Int, Mat4, Vec2, Vec3 } from "./values"
+
+/**
+ * Returns the current fragment's on-screen coordinate.
+ */
+export const FragmentCoordinate = Vec2($`gl_FragCoord.xy`, {
+  name: "Fragment Coordinate",
+  only: "fragment"
+})
+
+/**
+ * In instanced rendering, will return the instance ID.
+ * Wraps the `gl_InstanceID` GLSL built-in.
+ *
+ * Note: available in vertex shader only!
+ */
+export const InstanceID = Int($`gl_InstanceID`, {
+  name: "Instance ID",
+  only: "vertex"
+})
+
+/**
+ * Returns the verte ID.
+ * Wraps the `gl_VertexID` GLSL built-in.
+ *
+ * Note: available in vertex shader only!
+ */
+export const VertexID = Int($`gl_VertexID`, {
+  name: "Vertex ID",
+  only: "vertex"
+})
 
 export const CameraPosition = Vec3($`cameraPosition`, {
   name: "Camera Position"
